requirements.txt
streamlit>=1.20.0
opencv-python
numpy
pandas
scikit-learn
fpdf
insightface
Pillow
onnxruntime
